% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_module_scores.R
\name{calc_background_sampling}
\alias{calc_background_sampling}
\title{Calculate Module/Pathway background, to be used}
\usage{
calc_background_sampling(mat, precal, nsample = 100, nrep = 20)
}
\arguments{
\item{mat}{expression matrix}

\item{precal}{gene bin vector to use specific order}

\item{nsample}{number of genes to sample for each target gene}

\item{nrep}{number of repeats to calculate sampling of bins}
}
\value{
matrix of background bin reps x cells
}
\description{
Calculate Module/Pathway background, to be used
}
