% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calc_modulescore}
\alias{calc_modulescore}
\title{Return string \code{"Hello world!"} to R.}
\usage{
calc_modulescore(mat, features, allfeatures, nbin, nsample, nthread)
}
\description{
Return string \code{"Hello world!"} to R.
}
