% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calc_modulescore_orderin}
\alias{calc_modulescore_orderin}
\title{Calculate pathway scoring, similar to Seurat::AddModuleScore, with precalculated order}
\usage{
calc_modulescore_orderin(
  mat,
  features,
  allfeatures,
  nbin,
  nsample,
  nthread,
  order
)
}
\description{
Calculate pathway scoring, similar to Seurat::AddModuleScore, with precalculated order
}
