% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmt_to_list}
\alias{gmt_to_list}
\title{Parse GMT file to gene lists}
\usage{
gmt_to_list(path, sep = "\\thttp.*?\\t", rm = "^.+?_")
}
\arguments{
\item{path}{gmt file (from http://www.gsea-msigdb.org/gsea/downloads.jsp)}

\item{sep}{remove string to keep path and genes}

\item{rm}{remove string from path name}
}
\value{
named list of genes
}
\description{
Parse GMT file to gene lists
}
