% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{order_expr}
\alias{order_expr}
\title{order genes by expression, random breaking ties}
\usage{
order_expr(mat, allfeatures, nthread)
}
\description{
order genes by expression, random breaking ties
}
