% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_module_scores.R
\name{order_expr_r}
\alias{order_expr_r}
\title{Order genes by overall expression}
\usage{
order_expr_r(mat, gs)
}
\arguments{
\item{mat}{expression matrix}
}
\value{
vector of gene names, high to low
}
\description{
Order genes by overall expression
}
