% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_module_scores.R, R/utils.R
\name{calc_module_scores}
\alias{calc_module_scores}
\title{Calculate Module/Pathway Scores , Parallel on R side}
\usage{
calc_module_scores(mat, paths, nbin = 24, nsample = 100, nthread = 4)

calc_module_scores(mat, paths, nbin = 24, nsample = 100, nthread = 4)
}
\arguments{
\item{mat}{expression matrix}

\item{paths}{named lists, each with genes of one pathway}

\item{nbin}{number of bins to arrange all genes}

\item{nsample}{number of genes to sample for each target gene}

\item{nthread}{number of threads for parallel processing of many pathways}

\item{precal}{option on how background order is calculated, set to TRUE to use same calculations for every pathway, pass vector to use specific order}

\item{nrep}{number of repeats to calculate sampling of bins}

\item{gmt}{GMT file}

\item{rm}{characters to remove from entry names}

\item{nmin}{minimum number of genes to keep pathway}
}
\value{
matrix of pathways x cells

named list of vectors, each representing a pathway
}
\description{
Calculate Module/Pathway Scores , Parallel on R side

Parse GMT format to get list of pathways
}
