% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calc_modulescore}
\alias{calc_modulescore}
\title{Calculate pathway scoring, similar to Seurat::AddModuleScore}
\usage{
calc_modulescore(mat, features, allfeatures, nbin, nsample, nthread)
}
\description{
Calculate pathway scoring, similar to Seurat::AddModuleScore
}
